Support for splitting out single data files into multiple epochs of observations, since this is something I want to do a lot.
Make a few more commands accept filters optionally.

Have sc.sh check for existence of calfluxes.tab before
attemping to run selfcals.



git-svn-id: https://svn.hcro.org/mmm@133 9b194dde-2b4d-0410-98f0-bba1668eb47e

+++ pwilliams/ws/icrf.sh
@@ -4,40 +4,46 @@
 
 . common.sh
 
-srcs=`./index-list.py mf.list src "$@"`
-freqs=`./index-list.py mf.list freq "$@"`
+srcs=`./index-list.py sc.list src "$@"`
+freqs=`./index-list.py sc.list freq "$@"`
+epochs=`./index-list.py sc.list epoch "$@"`
 flux=""
 
 for src in $srcs ; do
     for freq in $freqs ; do
-	vises=$(./comma.py $(./filter-list.py mf.list src=$src freq=$freq "$@"))
+	for epoch in $epochs ; do
+	    vises=$(./comma.py $(./filter-list.py sc.list \
+		src=$src freq=$freq epoch=$epoch "$@"))
 
-	echo $vises ...
+	    [ x"$vises" = x ] && continue
+	
+	    echo $vises ...
 
-	mp=mp-$src-$freq
-	bm=bm-$src-$freq
-	cl=cl-$src-$freq
-	rm=rm-$src-$freq
-	rs=rs-$src-$freq
+	    mp=mp-$src-$freq-$epoch
+	    bm=bm-$src-$freq-$epoch
+	    cl=cl-$src-$freq-$epoch
+	    rm=rm-$src-$freq-$epoch
+	    rs=rs-$src-$freq-$epoch
 
-	rm -rf $mp $bm $cl $rm $rs
-	cmd invert vis=$vises map=$mp beam=$bm options=double,mfs sup=0
-	cmd clean map=$mp beam=$bm out=$cl niters=5000
-	cmd restor map=$mp beam=$bm model=$cl out=$rm
-
-	# Imfit. We want to preserve the output so things get
-	# messier
-
-	imfit region='arcsec,box(-500,-500,500,500)' object=point \
-	    options=resid in=$rm out=$rs >$rm/ws-imfit
-	cat $rm/ws-imfit >>wank.log
-	pk=`grep 'Peak' $rm/ws-imfit`
-	model=`cat calfluxes.tab |grep "^$src $freq" |cut -d' ' -f 3`
-	flux="$flux
-$src $freq:
+	    rm -rf $mp $bm $cl $rm $rs
+	    cmd invert vis=$vises map=$mp beam=$bm options=double,mfs sup=0
+	    cmd clean map=$mp beam=$bm out=$cl niters=5000
+	    cmd restor map=$mp beam=$bm model=$cl out=$rm
+	    
+	    # Imfit. We want to preserve the output so things get
+	    # messier
+	    
+	    imfit region='arcsec,box(-500,-500,500,500)' object=point \
+		options=resid in=$rm out=$rs >$rm/ws-imfit
+	    cat $rm/ws-imfit >>wank.log
+	    pk=`grep 'Peak' $rm/ws-imfit`
+	    model=`cat calfluxes.tab |grep "^$src $freq" |cut -d' ' -f 3`
+	    flux="$flux
+$src $freq $epoch:
 $pk
   Model: $model
 "
+	done
     done
 done
